feat(railgun): introduce privacy-aware Recipient struct and extract core types
- Add `Recipient` struct (`name`, `evm_address`, `zk_address`) with `from_contact`, `from_wallet_info`, `from_unknown_evm_address`, `from_unknown_zk_address` constructors and `is_empty(privacy_mode)`
- Refactor `RecipientSelectionWindow` to store/return `Recipient` (instead of raw `String` + separate `recipient_name`)
- Update all recipient flows (send_crypto, railgun/shield, across) to use privacy-aware `.evm_address`/`.zk_address` and `is_empty(privacy_mode)` checks
- Create `src/core/types.rs` and move `Contact` (now zk-aware), `Dapp`, `BaseFee`, `PriorityFee`, `Block`, `EthCall`, `EstimateGas`, `ConnectedDapps`, `DelegatedWallets` etc. into the new module
- Remove the duplicated structs from `ctx.rs` (large cleanup) and update `core/mod.rs`; adjust imports across tx, gui, tests, and utils
