+++ src/core/mod.rs
@@ -1,6 +1,7 @@
 pub mod context;
 pub mod signature;
 pub mod tx;
+pub mod types;
 pub mod vault;
 pub mod wallet;
 
@@ -10,21 +11,6 @@ pub use tx::{analysis::*, events::*, rich::*, send::*};
 pub use vault::*;
 pub use wallet::*;
 
-#[derive(Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
-pub enum Dapp {
-   Across,
-   Uniswap,
-}
-
-impl Dapp {
-   pub fn is_across(&self) -> bool {
-      matches!(self, Self::Across)
-   }
-
-   pub fn is_uniswap(&self) -> bool {
-      matches!(self, Self::Uniswap)
-   }
-}
 
 mod serde_hashmap {
    use serde::{Deserialize, Deserializer, Serialize, Serializer, de::DeserializeOwned};
