+++ src/core/tx/analysis.rs
@@ -1,4 +1,4 @@
-use crate::core::{Dapp, ZeusCtx};
+use crate::core::{types::Dapp, ZeusCtx};
 use crate::utils::truncate_address;
 use alloy_eips::eip7702::SignedAuthorization;
 use serde::{Deserialize, Serialize};
