+++ src/gui/ui/dapps/across.rs
@@ -176,12 +176,11 @@ impl AcrossBridge {
 
       recipient_selection.show(ctx.clone(), theme, icons.clone(), false, contacts_ui, ui);
       let recipient = recipient_selection.get_recipient();
-      let recipient_name = recipient_selection.get_recipient_name();
       let from_chain = self.from_chain.chain.id();
       let depositor = ctx.current_wallet_info(false).address;
       self.currency = NativeCurrency::from(from_chain).into();
 
-      self.get_suggested_fees(ctx.clone(), depositor, &recipient);
+      self.get_suggested_fees(ctx.clone(), depositor, &recipient.evm_address);
 
       let frame = theme.frame1;
       let button_visuals = theme.button_visuals();
@@ -280,8 +279,8 @@ impl AcrossBridge {
                      ui.label(RichText::new("Recipient").size(theme.text_sizes.large));
                      ui.add_space(10.0);
 
-                     if !recipient.is_empty() {
-                        if let Some(name) = &recipient_name {
+                     if !recipient.is_empty(false) {
+                        if let Some(name) = &recipient.name {
                            ui.label(
                               RichText::new(name)
                                  .size(theme.text_sizes.large)
@@ -299,7 +298,7 @@ impl AcrossBridge {
 
                         let chain = self.to_chain.chain;
                         let block_explorer = chain.block_explorer();
-                        let link = format!("{}/address/{}", block_explorer, recipient);
+                        let link = format!("{}/address/{}", block_explorer, recipient.evm_address);
                         let icon = match theme.dark_mode {
                            true => icons.external_link_white_x18(tint),
                            false => icons.external_link_dark_x18(tint),
@@ -321,7 +320,7 @@ impl AcrossBridge {
                         .color(theme.colors.text_muted);
 
                      let res = ui.add(
-                        SecureTextEdit::singleline(&mut recipient_selection.recipient)
+                        SecureTextEdit::singleline(&mut recipient_selection.recipient.evm_address)
                            .visuals(visuals)
                            .hint_text(hint)
                            .min_size(vec2(ui_width, 25.0))
@@ -405,7 +404,7 @@ impl AcrossBridge {
                   }
                });
 
-               self.bridge_button(ctx.clone(), theme, depositor, recipient, ui);
+               self.bridge_button(ctx.clone(), theme, depositor, recipient.evm_address, ui);
             });
          });
    }
