+++ src/gui/ui/settings/contacts.rs
@@ -1,7 +1,7 @@
 //! UI that allows the user to add,edit and remove contacts.
 
 use crate::assets::icons::Icons;
-use crate::core::{Contact, ZeusCtx};
+use crate::core::{types::Contact, ZeusCtx};
 use crate::gui::SHARED_GUI;
 use crate::utils::RT;
 use egui::{
