+++ src/tests/swap.rs
@@ -1,6 +1,6 @@
 #[cfg(test)]
 mod tests {
-   use crate::core::{BaseFee, ZeusCtx};
+   use crate::core::{types::BaseFee, ZeusCtx};
    use crate::gui::ui::dapps::uniswap::swap::get_relevant_pools;
 
    use crate::utils::{swap_quoter::*, universal_router_v2::*};
